user_tasks/tasks/serializers.py
# tasks/serializers.py

from rest_framework import serializers
from .models import Task, Comment

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['title', 'description', 'due_date', 'status', 'created_by']

class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = '__all__'
