user_tasks/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2